+++ R/corr_feats.R
@@ -2,6 +2,8 @@
 #' Identify correlated features
 #'
 #' @param features Features used for machine learning.
+#' @param corr_method correlation method. options or the same as those supported
+#'   by `stats::cor`: spearman, pearson, kendall. (default: spearman)
 #' @param group_neg_corr Whether to group negatively correlated features
 #'   together (e.g. c(0,1) and c(1,0)).
 #' @inheritParams run_ml
@@ -20,9 +22,10 @@
 #' colnames(mat) <- 1:ncol(mat)
 #' get_corr_feats(mat, 0.4)
 #' @importFrom dplyr .data
-get_corr_feats <- function(features, corr_thresh = 1, group_neg_corr = TRUE) {
+get_corr_feats <- function(features, corr_thresh = 1, group_neg_corr = TRUE,
+                           corr_method = "spearman") {
   corr_feats <- features %>%
-    stats::cor(method = "spearman") %>%
+    stats::cor(method = corr_method) %>%
     flatten_corr_mat()
   if (group_neg_corr) {
     corr_feats <- corr_feats %>%
@@ -62,7 +65,7 @@ flatten_corr_mat <- function(cormat) {
 #'
 #' @param features data frame with each column as a feature for ML
 #' @param corr_thresh correlation threshold (default: 1)
-#'
+#' @inheritParams get_corr_feats
 #' @return vector of correlated features where each element is a group of
 #'   correlated features separated by pipes (|)
 #' @noRd
@@ -70,10 +73,11 @@ flatten_corr_mat <- function(cormat) {
 #' @author Zena Lapp, \email{zenalapp@@umich.edu}
 #'
 group_correlated_features <- function(features, corr_thresh = 1,
-                                      group_neg_corr = TRUE) {
+                                      group_neg_corr = TRUE, corr_method = "spearman") {
   corr <- get_corr_feats(features,
     corr_thresh = corr_thresh,
-    group_neg_corr = group_neg_corr
+    group_neg_corr = group_neg_corr,
+    corr_method = corr_method
   )
   corr <- dplyr::select_if(corr, !(names(corr) %in% c("corr")))