+++ tests/testthat/test-corr_feats.R
@@ -30,6 +30,18 @@ test_that("get_corr_feats works", {
     c = c(0.6607978, 0.6291140, 0.06178627, 0.2059746)
   ), 0.6), cor_feats, tolerance = tol)
   expect_equivalent(get_corr_feats(data.frame(
+    a = c(0.3607978, 0.2911404, 0.36178627, 0.6059746),
+    b = c(0.2099978, 0.3013719, 0.59891969, 0.6442003),
+    c = c(0.9066972, 0.6655073, 0.37212390, 0.0728404)
+  ), 0.9, corr_method = "pearson"),
+  data.frame(
+    feature1 = c("b"),
+    feature2 = c("c"),
+    corr = c(-0.96)
+  ),
+  tolerance = tol
+  )
+  expect_equivalent(get_corr_feats(data.frame(
     a = c(1, 1, 0, 0),
     b = c(1, 1, 0, 0),
     c = c(0, 0, 1, 1)
@@ -58,6 +70,17 @@ test_that("group_correlated_features works", {
   expect_equal(
     sort(group_correlated_features(data.frame(a = 1:3, b = c(3, 1, 2)))),
     c("a", "b")
+  )
+  expect_equal(
+    sort(group_correlated_features(data.frame(
+      a = c(1, 0, 0),
+      b = c(3, 2, 4),
+      c = c(1, 3, 4)
+    ),
+    corr_thresh = 0.9,
+    corr_method = "pearson"
+    )),
+    c("b", "c|a")
   )
 
   corr <- dplyr::tibble(feature1 = c(character()), feature2 = character())